Fix Issue with Win Size Restore on Drop
If a window is moved less than the threshold, the rect is not being
restored to the size it was when it was originally dropped. This logic
may change in the future, but for now this fixes a bug in the behavior
which I believe is causing unusual behavior when moving windows.